Timer: Fix event timer lagging
Event timer (RX,TX,SQ) was lagging when comparing with wall time.
That was due to the inherent processing time plus the wait(1000),
which builds up over the time. Now the events timer is relative to
the uptime timer, which has a better precision and not subjected
to processing times.